backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cmake==3.31.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
cvzone==1.6.1
deep-sort-realtime==1.3.2
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
dlib @ file:///C:/Users/muhsi/Downloads/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dnspython==2.7.0
email_validator==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
filterpy==1.4.5
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
git-filter-repo==2.47.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jedi==0.17.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Markdown==3.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mongoengine==0.29.1
mpmath==1.3.0
natsort==8.4.0
networkx==3.5
numpy==2.2.5
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
parso==0.7.1
pillow==11.3.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
psutil==7.0.0
psycopg==3.2.5
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.1
pyotp==2.9.0
pyparsing==3.2.3
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
scipy==1.16.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
twilio==9.6.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==14.2
wheel==0.45.1
whitenoise==6.9.0
yarl==1.20.1
